requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
motor
aiohttp
requests
openai
python-multipart